requirements.txt
numpy
opencv_python
opencv_python_headless
Pillow
torch
torchvision
scikit-image